--- bin/afj-rest.js
@@ -1,6 +1,0 @@
-#!/usr/bin/env node
-/* eslint-disable @typescript-eslint/no-var-requires, no-undef */
-
-import { runCliServer } from '../build/cli.js'
-
-runCliServer()
