--- schema/schemaAttributes.json
@@ -1,13 +1,0 @@
-{
-    "issuerId": "did:key:z6MkrDn3MqmedCnj4UPBwZ7nLTBmK9T9BwB3njFmQRUqoFn1",
-    "version": "1.0", 
-    "name":"string",
-    "attrNames": [ 
-        "checkName",
-        "companyName" ,
-        "companiesHouseNumber",
-        "issueDate",
-        "expiryDate"]
-    
-  }
-
