+++ src/controllers/v1/agent/AgentController.ts
@@ -3,11 +3,12 @@ import { Controller, Get, Route, Tags } from 'tsoa'
 import { injectable } from 'tsyringe'
 
 import type { AgentInfo } from '../../types.js'
+import { RestAgent } from '../../../agent.js'
 @Tags('Agent')
 @Route('/v1/agent')
 @injectable()
 export class AgentController extends Controller {
-  private agent: Agent
+  private agent: RestAgent
 
   public constructor(agent: Agent) {
     super()
