+++ src/controllers/v1/credentials/CredentialDefinitionController.ts
@@ -2,10 +2,10 @@ import { Agent } from '@credo-ts/core'
 import { Body, Controller, Example, Get, Path, Post, Route, Tags, Response, Query } from 'tsoa'
 import { injectable } from 'tsyringe'
 
-import type { RestAgent } from '../../../utils/agent.js'
 import { type Did, type SchemaId, type CredentialDefinitionId, CredentialDefinitionExample } from '../../examples.js'
 import type { AnonCredsCredentialDefinitionResponse } from '../../types.js'
 import { HttpResponse, NotFound, BadRequest } from '../../../error.js'
+import { RestAgent } from '../../../agent.js'
 
 @Tags('Credential Definitions')
 @Route('/v1/credential-definitions')
