+++ src/controllers/v1/credentials/SchemaController.ts
@@ -2,10 +2,10 @@ import { Agent } from '@credo-ts/core'
 import { Body, Example, Get, Path, Post, Route, Tags, Response, Query } from 'tsoa'
 import { injectable } from 'tsyringe'
 
-import type { RestAgent } from '../../../utils/agent.js'
 import { type Did, type Version, type SchemaId, SchemaExample } from '../../examples.js'
 import type { AnonCredsSchemaResponse } from '../../types.js'
 import { HttpResponse, NotFound, BadRequest } from '../../../error.js'
+import { RestAgent } from '../../../agent.js'
 
 @Tags('Schemas')
 @Route('/v1/schemas')
