+++ src/policyAgent/index.ts
@@ -1,6 +1,7 @@
 import { singleton } from 'tsyringe'
 
 import { HttpResponse, NotFound } from '../error.js'
+import { Env } from '../env.js'
 
 type Policy = {
   id: string
@@ -14,11 +15,14 @@ type Policy = {
 }
 @singleton()
 export default class PolicyAgent {
-  constructor(private origin: string) {
+  private origin: string
+
+  constructor(private env: Env) {
+    this.origin = env.get('OPA_ORIGIN')
     try {
-      new URL(origin)
+      new URL(this.origin)
     } catch (err) {
-      throw new Error(`Invalid PolicyAgent origin ${origin}`)
+      throw new Error(`Invalid PolicyAgent origin ${this.origin}`)
     }
   }
