+++ src/server.ts
@@ -10,7 +10,7 @@ import path from 'path'
 import { fileURLToPath } from 'url'
 
 import type { ServerConfig } from './utils/ServerConfig.js'
-import type { RestAgent } from './utils/agent.js'
+
 import { basicMessageEvents } from './events/BasicMessageEvents.js'
 import { connectionEvents } from './events/ConnectionEvents.js'
 import { credentialEvents } from './events/CredentialEvents.js'
@@ -18,10 +18,9 @@ import { proofEvents } from './events/ProofEvents.js'
 import { trustPingEvents } from './events/TrustPingEvents.js'
 import { drpcEvents } from './events/DrpcEvents.js'
 import { verifiedDrpcEvents } from './events/VerifiedDrpcEvents.js'
-//import { VerifiedDrpcService } from './verified-drpc/VerifiedDrpcService.js'
 import { RegisterRoutes } from './routes/routes.js'
 import { errorHandler } from './error.js'
-import PolicyAgent from './policyAgent/index.js'
+import { RestAgent } from './agent.js'
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = path.dirname(__filename)
@@ -31,9 +30,8 @@ export const setupServer = async (agent: RestAgent, config: ServerConfig) => {
   const swaggerJson = JSON.parse(swaggerBuffer.toString('utf8'))
 
   container.registerInstance(Agent, agent as Agent)
-  container.registerInstance(PolicyAgent, new PolicyAgent(config.opaOrigin || 'http://localhost:8181'))
 
-  const app = config.app ?? express()
+  const app = express()
   if (config.cors) app.use(cors())
 
   if (config.socketServer || (config.webhookUrl && config.webhookUrl.length > 0)) {
