+++ tests/unit/basicMessage.test.ts
@@ -4,16 +4,16 @@ import { expect } from 'chai'
 import { spy, restore as sinonRestore } from 'sinon'
 
 import type { Agent, BasicMessageRecord, ConnectionRecord } from '@credo-ts/core'
-import type { Server } from 'net'
 
+import { AddressInfo, Server } from 'node:net'
 import { BasicMessageEventTypes } from '@credo-ts/core'
 import request from 'supertest'
 import WebSocket from 'ws'
 
-import { startServer } from '../../src/index.js'
-import { getTestAgent, objectToJson } from './utils/helpers.js'
+import { getTestAgent, getTestServer, objectToJson } from './utils/helpers.js'
 
 describe('BasicMessageController', () => {
+  let port: number
   let server: Server
   let aliceAgent: Agent
   let bobAgent: Agent
@@ -22,7 +22,8 @@ describe('BasicMessageController', () => {
   before(async () => {
     aliceAgent = await getTestAgent('Basic Message REST Agent Test Alice', 3002)
     bobAgent = await getTestAgent('Basic Message REST Agent Test Bob', 5034)
-    server = await startServer(bobAgent, { port: 5033 })
+    server = await getTestServer(bobAgent)
+    port = (server.address() as AddressInfo).port
 
     const { outOfBandInvitation } = await aliceAgent.oob.createInvitation()
     const { outOfBandRecord: bobOOBRecord } = await bobAgent.oob.receiveInvitation(outOfBandInvitation)
@@ -55,7 +56,7 @@ describe('BasicMessageController', () => {
 
   describe('Basic Message WebSocket event', () => {
     test('should return basic message event sent from test agent to clients', async () => {
-      const client = new WebSocket('ws://localhost:5033')
+      const client = new WebSocket(`ws://localhost:${port}`)
 
       const waitForMessagePromise = new Promise((resolve) => {
         client.on('message', (data) => {
