+++ tests/unit/credential.test.ts
@@ -10,7 +10,7 @@ import type {
   GetCredentialFormatDataReturn,
   OutOfBandRecord,
 } from '@credo-ts/core'
-import type { Server } from 'net'
+import type { AddressInfo, Server } from 'node:net'
 
 import {
   AutoAcceptCredential,
@@ -26,7 +26,7 @@ import {
 import request from 'supertest'
 import WebSocket from 'ws'
 
-import { startServer } from '../../src/index.js'
+import { setupServer } from '../../src/server.js'
 
 import {
   objectToJson,
@@ -36,10 +36,12 @@ import {
   getTestOutOfBandRecord,
   getTestConnection,
   getCredentialFormatData,
+  getTestServer,
 } from './utils/helpers.js'
 import { AnonCredsCredentialFormat } from '@credo-ts/anoncreds'
 
 describe('CredentialController', () => {
+  let port: number
   let app: Server
   let aliceAgent: Agent
   let bobAgent: Agent
@@ -55,7 +57,8 @@ describe('CredentialController', () => {
   before(async () => {
     aliceAgent = await getTestAgent('Credential REST Agent Test Alice', 3022)
     bobAgent = await getTestAgent('Credential REST Agent Test Bob', 3023)
-    app = await startServer(bobAgent, { port: 3024 })
+    app = await getTestServer(bobAgent)
+    port = (app.address() as AddressInfo).port
 
     testCredential = getTestCredential() as CredentialExchangeRecord
     testFormatData = getCredentialFormatData()
@@ -307,7 +310,7 @@ describe('CredentialController', () => {
       const now = new Date()
 
       // Start client and wait for it to be opened
-      const client = new WebSocket('ws://localhost:3024')
+      const client = new WebSocket(`ws://localhost:${port}`)
       await new Promise((resolve) => client.once('open', resolve))
 
       // Start promise to listen for message
