+++ tests/unit/proof.test.ts
@@ -5,21 +5,22 @@ import type {
   RequestProofOptions,
 } from '../../src/controllers/types.js'
 import type { Agent, ProofStateChangedEvent } from '@credo-ts/core'
-import type { Server } from 'net'
+import type { AddressInfo, Server } from 'node:net'
 
-import { describe, before, beforeEach, after, afterEach, test } from 'mocha'
-import { expect, use as chaiUse, Assertion as assertion } from 'chai'
+import { describe, before, after, afterEach, test } from 'mocha'
+import { expect } from 'chai'
 import { stub, restore as sinonRestore } from 'sinon'
 
 import { AgentMessage, ProofEventTypes, ProofExchangeRecord, ProofRole, ProofState } from '@credo-ts/core'
 import request from 'supertest'
 import WebSocket from 'ws'
 
-import { startServer } from '../../src/index.js'
+import { setupServer } from '../../src/server.js'
 
-import { getTestAgent, getTestProof, getTestProofResponse, objectToJson } from './utils/helpers.js'
+import { getTestAgent, getTestProof, getTestProofResponse, getTestServer, objectToJson } from './utils/helpers.js'
 
 describe('ProofController', () => {
+  let port: number
   let app: Server
   let aliceAgent: Agent
   let bobAgent: Agent
@@ -30,7 +31,8 @@ describe('ProofController', () => {
   before(async () => {
     aliceAgent = await getTestAgent('Proof REST Agent Test Alice', 3032)
     bobAgent = await getTestAgent('Proof REST Agent Test Bob', 3912)
-    app = await startServer(bobAgent, { port: 3033 })
+    app = await getTestServer(bobAgent)
+    port = (app.address() as AddressInfo).port
 
     testProof = getTestProof()
     testProofResponse = getTestProofResponse()
@@ -363,7 +365,7 @@ describe('ProofController', () => {
       })
 
       // Start client and wait for it to be opened
-      const client = new WebSocket('ws://localhost:3033')
+      const client = new WebSocket(`ws://localhost:${port}`)
       await new Promise((resolve) => client.once('open', resolve))
 
       // Start promise to listen for message
