+++ tests/unit/schema.test.ts
@@ -1,6 +1,6 @@
-import type { RestAgent } from '../../src/utils/agent.js'
+import type { RestAgent } from '../../src/agent.js'
 import type { AnonCredsSchema, AnonCredsSchemaRecord } from '@credo-ts/anoncreds'
-import type { Express } from 'express'
+import type { Server } from 'node:net'
 
 import { describe, before, after, afterEach, test } from 'mocha'
 import { expect } from 'chai'
@@ -10,18 +10,17 @@ import request from 'supertest'
 
 import { setupServer } from '../../src/server.js'
 
-import { getTestAgent, getTestSchema } from './utils/helpers.js'
-import _schema from '../../schema/schemaAttributes.json'
-const schema = _schema as AnonCredsSchema
+import { getTestAgent, getTestSchema, getTestServer } from './utils/helpers.js'
+import { schema } from './utils/fixtures.js'
 
 describe('SchemaController', () => {
-  let app: Express
+  let app: Server
   let agent: RestAgent
   let testSchema: AnonCredsSchema
 
   before(async () => {
     agent = await getTestAgent('Schema REST Agent Test', 3021)
-    app = await setupServer(agent, { port: 3000 })
+    app = await getTestServer(agent)
     testSchema = getTestSchema()
   })
 
@@ -222,5 +221,6 @@ describe('SchemaController', () => {
   after(async () => {
     await agent.shutdown()
     await agent.wallet.delete()
+    app.close()
   })
 })
