+++ tests/unit/utils/fixtures.ts
@@ -1,0 +1,6 @@
+export const schema = {
+  issuerId: 'did:key:z6MkrDn3MqmedCnj4UPBwZ7nLTBmK9T9BwB3njFmQRUqoFn1',
+  version: '1.0',
+  name: 'string',
+  attrNames: ['checkName', 'companyName', 'companiesHouseNumber', 'issueDate', 'expiryDate'],
+}
