+++ tsoa.json
@@ -8,7 +8,7 @@
   },
   "routes": {
     "routesDir": "src/routes",
-    "iocModule": "./src/utils/tsyringeTsoaIocContainer",
+    "iocModule": "./src/ioc",
     "esm": true
   }
 }
